my_review_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "my-review-agent"
version = "1.0.0"
description = "A Gemini-powered agent to perform systematic searches on NCBI databases like PubMed."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "google-generativeai>=0.5.0",
    "biopython>=1.83",
    "pandas>=2.0",
    "xmltodict>=0.13.0"
]

# A command-line script to run the application
[project.scripts]
run-agent = "my_review_agent.app:run"
